Hash works and is test covered to 94%
I fixed the Hash to work and got it test covered to 94%. I also got the style points back up to 20/20 so now we are only getting flagged for TODO comments. 
Next task is to fix up the BST (if thats the next piece to be fixed up) and implement the Handle class somewhere in there.